+++ zhalilov.rail/S4/const_treeIterator.hpp
@@ -1,50 +1,48 @@
-#ifndef ITERATOR_HPP
-#define ITERATOR_HPP
+#ifndef CONST_TREEITERATOR_HPP
+#define CONST_TREEITERATOR_HPP
 
 #include <iterator>
 
-#include "node.hpp"
+#include "treeNode.hpp"
 
 namespace zhalilov
 {
   template < class Key, class T, class Compare >
-  class TwoThreeTree < Key, T, Compare >;
+  class TwoThree;
 
   template < class T >
-  class TwoThreeIterator: public std::iterator < std::bidirectional_iterator_tag, T >
+  class ConstTwoThreeIterator: public std::iterator < std::bidirectional_iterator_tag, const T >
   {
   public:
-    ~TwoThreeIterator() = default;
+    ~ConstTwoThreeIterator() = default;
 
-    TwoThreeIterator &operator=(const TwoThreeIterator &) = default;
+    ConstTwoThreeIterator &operator=(const ConstTwoThreeIterator &) = default;
 
-    TwoThreeIterator &operator++();
-    TwoThreeIterator &operator--();
-    TwoThreeIterator operator++(int);
-    TwoThreeIterator operator--(int);
+    ConstTwoThreeIterator &operator++();
+    ConstTwoThreeIterator &operator--();
+    ConstTwoThreeIterator operator++(int);
+    ConstTwoThreeIterator operator--(int);
 
-    T &operator*();
-    T *operator->();
     const T &operator*() const;
     const T *operator->() const;
 
-    bool operator==(const TwoThreeIterator &) const;
-    bool operator!=(const TwoThreeIterator &) const;
+    bool operator==(const ConstTwoThreeIterator &) const;
+    bool operator!=(const ConstTwoThreeIterator &) const;
 
     template < class Key, class Value, class Compare >
-    friend TwoThreeTree < Key, Value, Compare >;
+    friend TwoThree < Key, Value, Compare >;
 
   private:
     detail::Node < T > *node_;
     bool isPtrToLeft_;
 
-    explicit TwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft = true);
+    explicit ConstTwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft = true);
     static detail::Node < T > *findMin(const detail::Node < T > *);
     static detail::Node < T > *findMax(const detail::Node < T > *);
   };
 
   template < class T >
-  TwoThreeIterator < T > &TwoThreeIterator < T >::operator++()
+  ConstTwoThreeIterator < T > &ConstTwoThreeIterator < T >::operator++()
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -78,7 +76,7 @@ namespace zhalilov
   }
 
   template < class T >
-  TwoThreeIterator < T > &TwoThreeIterator < T >::operator--()
+  ConstTwoThreeIterator < T > &ConstTwoThreeIterator < T >::operator--()
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -112,23 +110,23 @@ namespace zhalilov
   }
 
   template < class T >
-  TwoThreeIterator < T > TwoThreeIterator < T >::operator++(int)
+  ConstTwoThreeIterator < T > ConstTwoThreeIterator < T >::operator++(int)
   {
-    TwoThreeIterator temp(*this);
+    ConstTwoThreeIterator temp(*this);
     operator++();
     return temp;
   }
 
   template < class T >
-  TwoThreeIterator < T > TwoThreeIterator < T >::operator--(int)
+  ConstTwoThreeIterator < T > ConstTwoThreeIterator < T >::operator--(int)
   {
-    TwoThreeIterator temp(*this);
+    ConstTwoThreeIterator temp(*this);
     operator--();
     return temp;
   }
 
   template < class T >
-  T &TwoThreeIterator < T >::operator*()
+  const T &ConstTwoThreeIterator < T >::operator*() const
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -141,7 +139,7 @@ namespace zhalilov
   }
 
   template < class T >
-  T *TwoThreeIterator < T >::operator->()
+  const T *ConstTwoThreeIterator < T >::operator->() const
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -154,19 +152,7 @@ namespace zhalilov
   }
 
   template < class T >
-  const T &TwoThreeIterator < T >::operator*() const
-  {
-    return operator*();
-  }
-
-  template < class T >
-  const T *TwoThreeIterator < T >::operator->() const
-  {
-    return operator->();
-  }
-
-  template < class T >
-  bool TwoThreeIterator < T >::operator==(const TwoThreeIterator < T > &ait) const
+  bool ConstTwoThreeIterator < T >::operator==(const ConstTwoThreeIterator < T > &ait) const
   {
     bool isEqualValue = true;
     if (node_->type == detail::NodeType::Three)
@@ -177,19 +163,19 @@ namespace zhalilov
   }
 
   template < class T >
-  bool TwoThreeIterator < T >::operator!=(const TwoThreeIterator < T > &ait) const
+  bool ConstTwoThreeIterator < T >::operator!=(const ConstTwoThreeIterator < T > &ait) const
   {
     return !operator==(ait);
   }
 
   template < class T >
-  TwoThreeIterator < T >::TwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft):
+  ConstTwoThreeIterator < T >::ConstTwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft):
     node_(node),
     isPtrToLeft_(isPtrToLeft)
   {}
 
   template < class T >
-  detail::Node < T > *TwoThreeIterator < T >::findMin(const detail::Node < T > *nodeFrom)
+  detail::Node < T > *ConstTwoThreeIterator < T >::findMin(const detail::Node < T > *nodeFrom)
   {
     detail::Node < T > *minNode = nodeFrom;
     while (minNode && minNode->left)
@@ -200,7 +186,7 @@ namespace zhalilov
   }
 
   template < class T >
-  detail::Node < T > *TwoThreeIterator < T >::findMax(const detail::Node < T > *nodeFrom)
+  detail::Node < T > *ConstTwoThreeIterator < T >::findMax(const detail::Node < T > *nodeFrom)
   {
     detail::Node < T > *maxNode = nodeFrom;
     while (maxNode && maxNode->right)
