+++ zhalilov.rail/S4/treeIterator.hpp
@@ -1,48 +1,50 @@
-#ifndef CONST_ITERATOR_HPP
-#define CONST_ITERATOR_HPP
+#ifndef TREEITERATOR_HPP
+#define TREEITERATOR_HPP
 
 #include <iterator>
 
-#include "node.hpp"
+#include "treeNode.hpp"
 
 namespace zhalilov
 {
   template < class Key, class T, class Compare >
-  class TwoThreeTree < Key, T, Compare >;
+  class TwoThreeTree;
 
   template < class T >
-  class ConstTwoThreeIterator: public std::iterator < std::bidirectional_iterator_tag, const T >
+  class TwoThreeIterator: public std::iterator < std::bidirectional_iterator_tag, T >
   {
   public:
-    ~ConstTwoThreeIterator() = default;
+    ~TwoThreeIterator() = default;
 
-    ConstTwoThreeIterator &operator=(const ConstTwoThreeIterator &) = default;
+    TwoThreeIterator &operator=(const TwoThreeIterator &) = default;
 
-    ConstTwoThreeIterator &operator++();
-    ConstTwoThreeIterator &operator--();
-    ConstTwoThreeIterator operator++(int);
-    ConstTwoThreeIterator operator--(int);
+    TwoThreeIterator &operator++();
+    TwoThreeIterator &operator--();
+    TwoThreeIterator operator++(int);
+    TwoThreeIterator operator--(int);
 
+    T &operator*();
+    T *operator->();
     const T &operator*() const;
     const T *operator->() const;
 
-    bool operator==(const ConstTwoThreeIterator &) const;
-    bool operator!=(const ConstTwoThreeIterator &) const;
+    bool operator==(const TwoThreeIterator &) const;
+    bool operator!=(const TwoThreeIterator &) const;
 
     template < class Key, class Value, class Compare >
-    friend TwoThreeTree < Key, Value, Compare >;
+    friend TwoThree < Key, Value, Compare >;
 
   private:
-    detail::Node < T > *node_;
+    detail::TreeNode < T > *node_;
     bool isPtrToLeft_;
 
-    explicit ConstTwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft = true);
+    explicit TwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft = true);
     static detail::Node < T > *findMin(const detail::Node < T > *);
     static detail::Node < T > *findMax(const detail::Node < T > *);
   };
 
   template < class T >
-  ConstTwoThreeIterator < T > &ConstTwoThreeIterator < T >::operator++()
+  TwoThreeIterator < T > &TwoThreeIterator < T >::operator++()
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -76,7 +78,7 @@ namespace zhalilov
   }
 
   template < class T >
-  ConstTwoThreeIterator < T > &ConstTwoThreeIterator < T >::operator--()
+  TwoThreeIterator < T > &TwoThreeIterator < T >::operator--()
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -110,23 +112,23 @@ namespace zhalilov
   }
 
   template < class T >
-  ConstTwoThreeIterator < T > ConstTwoThreeIterator < T >::operator++(int)
+  TwoThreeIterator < T > TwoThreeIterator < T >::operator++(int)
   {
-    ConstTwoThreeIterator temp(*this);
+    TwoThreeIterator temp(*this);
     operator++();
     return temp;
   }
 
   template < class T >
-  ConstTwoThreeIterator < T > ConstTwoThreeIterator < T >::operator--(int)
+  TwoThreeIterator < T > TwoThreeIterator < T >::operator--(int)
   {
-    ConstTwoThreeIterator temp(*this);
+    TwoThreeIterator temp(*this);
     operator--();
     return temp;
   }
 
   template < class T >
-  const T &ConstTwoThreeIterator < T >::operator*() const
+  T &TwoThreeIterator < T >::operator*()
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -139,7 +141,7 @@ namespace zhalilov
   }
 
   template < class T >
-  const T *ConstTwoThreeIterator < T >::operator->() const
+  T *TwoThreeIterator < T >::operator->()
   {
     if (node_->type == detail::NodeType::Three)
     {
@@ -152,7 +154,19 @@ namespace zhalilov
   }
 
   template < class T >
-  bool ConstTwoThreeIterator < T >::operator==(const ConstTwoThreeIterator < T > &ait) const
+  const T &TwoThreeIterator < T >::operator*() const
+  {
+    return operator*();
+  }
+
+  template < class T >
+  const T *TwoThreeIterator < T >::operator->() const
+  {
+    return operator->();
+  }
+
+  template < class T >
+  bool TwoThreeIterator < T >::operator==(const TwoThreeIterator < T > &ait) const
   {
     bool isEqualValue = true;
     if (node_->type == detail::NodeType::Three)
@@ -163,19 +177,19 @@ namespace zhalilov
   }
 
   template < class T >
-  bool ConstTwoThreeIterator < T >::operator!=(const ConstTwoThreeIterator < T > &ait) const
+  bool TwoThreeIterator < T >::operator!=(const TwoThreeIterator < T > &ait) const
   {
     return !operator==(ait);
   }
 
   template < class T >
-  ConstTwoThreeIterator < T >::ConstTwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft):
+  TwoThreeIterator < T >::TwoThreeIterator(detail::Node < T > *node, bool isPtrToLeft):
     node_(node),
     isPtrToLeft_(isPtrToLeft)
   {}
 
   template < class T >
-  detail::Node < T > *ConstTwoThreeIterator < T >::findMin(const detail::Node < T > *nodeFrom)
+  detail::Node < T > *TwoThreeIterator < T >::findMin(const detail::Node < T > *nodeFrom)
   {
     detail::Node < T > *minNode = nodeFrom;
     while (minNode && minNode->left)
@@ -186,7 +200,7 @@ namespace zhalilov
   }
 
   template < class T >
-  detail::Node < T > *ConstTwoThreeIterator < T >::findMax(const detail::Node < T > *nodeFrom)
+  detail::Node < T > *TwoThreeIterator < T >::findMax(const detail::Node < T > *nodeFrom)
   {
     detail::Node < T > *maxNode = nodeFrom;
     while (maxNode && maxNode->right)
